improve findAvailablePort(); cleanup all running containers on exit
The old way only looked at ports created by botblitz in the current session, meaning that if there were containers running from a previous session or other apps on running on the computer using ports in the 8000-9000 range, an in-use port could be assigned, resulting in failure. The new method just asks the os for an unused port.